Tidy up note package
